+++ modules/database/main.tf
@@ -9,9 +9,8 @@ resource "random_string" "master_password" {
 }
 
 locals {
-  is_mysql_80            = length(regexall("^8.0", var.engine_version)) > 0
-  engine_version_tag     = local.is_mysql_80 ? "80" : "57"
-  parameter_family       = local.is_mysql_80 ? "aurora-mysql8.0" : "aurora-mysql5.7"
+  engine_version_tag     = "80"
+  parameter_family       = "aurora-mysql8.0"
   parameter_group_name   = "${var.namespace}-aurora-db-${local.engine_version_tag}-parameter-group"
   parameter_cluster_name = "${var.namespace}-aurora-${local.engine_version_tag}-cluster-parameter-group"
 }