+++ modules/iam_role/variables.tf
@@ -1,0 +1,12 @@
+variable "namespace" {
+  type        = string
+  description = "The name prefix for all resources created."
+}
+
+variable "aws_iam_openid_connect_provider_url" {
+  type        = string
+}
+
+variable "yace_sa_name" {
+  type = string
+}