+++ outputs.tf
@@ -48,8 +48,14 @@ output "internal_app_port" {
 }
 
 output "kms_key_arn" {
-  value       = local.kms_key_arn
+  value       = local.default_kms_key
   description = "The Amazon Resource Name of the KMS key used to encrypt data at rest."
+}
+
+output "kms_clickhouse_key_arn" {
+  value       = local.clickhouse_kms_key
+  description = "The Amazon Resource Name of the KMS key used to encrypt Weave data at rest in Clickhouse."
+
 }
 
 output "network_id" {